refactor: implement ListEntriesFactory and FilterEntriesByFolderFactory
